feature: Initial vision check system done
